fix: 메뉴 상세, 리뷰 추가 화면 수정
AppString 추가, Widget 따로저장, 리뷰버튼 회색(별점만 입력 안했을 시), 별점 기본 0, 이미지 삭제 버튼 하얀색
